misc/challenges/Casino_Drawch/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==23.0.0
PyJWT==2.8.0